+++ Runtime/Scripts/Components/BlurImage.cs
@@ -1,5 +1,7 @@
 using ParkMinPackages.Foundation.Constants;
+using ParkMinPackages.UGUI.Blur.RendererFeatures;
 using Sirenix.OdinInspector;
+using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Rendering;
 using UnityEngine.Serialization;
@@ -36,15 +38,34 @@ namespace ParkMinPackages.UGUI.Blur.Components
 				}
 			}
 		}
+		internal static BlurImageSource FindExplicitSourceFor(Camera camera) {
+			for (int i = 0; i < _activeImages.Count; i++) {
+				BlurImage blurImage = _activeImages[i];
+				if (blurImage != null && blurImage.isActiveAndEnabled && blurImage._source != null && blurImage._source.isActiveAndEnabled && blurImage._source.CanRenderFor(camera)) {
+					return blurImage._source;
+				}
+			}
+			return null;
+		}
+
+		// - Public Properties -
+		public BlurImageSource Source
+		{
+			get { return _source; }
+			set { _source = value; }
+		}
 
 		// - Handler -
 		protected override void OnEnable() {
 			base.OnEnable();
+			_activeImages.Remove(this);
+			_activeImages.Add(this);
 			MigrateLegacyTintColor();
 			Canvas.preWillRenderCanvases += OnCanvasRendering;
 			graphic.SetMaterialDirty();
 		}
 		protected override void OnDisable() {
+			_activeImages.Remove(this);
 			Canvas.preWillRenderCanvases -= OnCanvasRendering;
 			CoreUtils.Destroy(_material);
 			_material = null;
@@ -73,6 +94,7 @@ namespace ParkMinPackages.UGUI.Blur.Components
 		// - Private & Protected -
 		[Header(Headers.Required)]
 		[SerializeField, Required] Material _materialTemplate;
+		[InfoBox("A different BlurImageSource is selected for the same Camera. Only the first registered Source will be used.", InfoMessageType.Warning, nameof(HasSourceConflict)), SerializeField, Required, Tooltip("The source used to generate this image's blur texture. No blur is rendered when omitted.")] BlurImageSource _source;
 
 		[SerializeField, HideInInspector, FormerlySerializedAs("_tintColor")] Color _legacyTintColor = new Color(-1f, -1f, -1f, -1f);
 		[SerializeField, HideInInspector] bool _imageColorMigrated;
@@ -80,6 +102,20 @@ namespace ParkMinPackages.UGUI.Blur.Components
 		Material _material;
 		Material _baseMaterial;
 
+		bool HasSourceConflict() {
+			if (_source == null || _source.TryGetRenderCamera(out Camera camera) == false) {
+				return false;
+			}
+
+			for (int i = 0; i < _activeImages.Count; i++) {
+				BlurImage blurImage = _activeImages[i];
+				BlurImageSource otherSource = blurImage == null ? null : blurImage._source;
+				if (blurImage != this && blurImage.isActiveAndEnabled && otherSource != null && otherSource != _source && otherSource.isActiveAndEnabled && otherSource.CanRenderFor(camera)) {
+					return true;
+				}
+			}
+			return false;
+		}
 		void MigrateLegacyTintColor() {
 			if (_imageColorMigrated) {
 				return;
@@ -93,11 +129,13 @@ namespace ParkMinPackages.UGUI.Blur.Components
 		void UpdateMaterialProperties() {
 			if (_material != null && graphic != null) {
 				_material.SetColor(_tintColorPropertyId, graphic.color);
-				_material.SetFloat(_opacityPropertyId, 1f);
+				bool sourceAvailable = _source != null && _source.IsReady();
+				_material.SetFloat(_opacityPropertyId, sourceAvailable ? 1f : 0f);
 			}
 		}
 
 		// - Private Statics -
+		static readonly List<BlurImage> _activeImages = new List<BlurImage>();
 		static readonly int _tintColorPropertyId = Shader.PropertyToID("_TintColor");
 		static readonly int _opacityPropertyId = Shader.PropertyToID("_Opacity");
 	}